application structure designed footer and header added css modules installed and used simple test created
