+++ src/screens/Home/parts/Footer/Footer.module.scss
@@ -1,0 +1,4 @@
+.footer {
+  display: flex;
+  justify-content: center;
+}
