+++ src/screens/Home/parts/Header/Header.module.scss
@@ -1,0 +1,4 @@
+.header {
+  display: flex;
+  justify-content: center;
+}
